Add CONFIG_H marco to all files include config.h
To make the compile progress faster when the config.h change, I remove
the --include config.h option, to let the unchanged file not to recompile.
This removal break the MATRIX_ROWS and MATRIX_COLS refrence in
many files. In order to make it work, I add CONFIG_H marco to these file,
The CONFIG_H marco indicate where the config.h file is located, then
the file which need to include config.h could use this marco to include.