Force maximum number of key of 6kro mode to 6 when nkro is enabled.
With the nkro and 6kro turn on at the same time, some key would lost in boot
protocal, which is caused by the wrong size of maximum key used in 6kro.
When nkro is enabled, the KEYBOARD_REPORT_KEYS set to 14, but in the
boot protocal, only 6 keys are acceptted, and those key after index 8
will lost.
This patch force the maximum number of key in 6kro window to 6 on nkro
enabled, which would fix this problem.